fix(makefile): reduce cppcheck execution time
There is a lot of ifdefs in the support header file msp430.h provided by
TI. This slows down cppcheck because it checks all the define combinations.
There is no need to check this file, so exclude it from the list of
cppcheck files. Suppress the error about missing system includes
(ignores standard include <>) and suppress errors about unmatched
suppressions as well as unused functions.

Also
* Break the cppcheck rule into variables
* Add -f flag to "clean" to return ok even if build/ doesn't exist