feat(drivers): implement IO handling
The pinmapping for the Launchpad and the actual PCB (Nsumo) are
represented with an enum (io_e). Create a second enum (io_generic_e),
define macros, and static register arrays (e.g. port_dir_regs) to map
the enum values to the corresponding registers in a space-efficient way.
Implement the individual functions to select function, set resistor,
direction, and output. The result is an abstraction that's much nicer
than directly accessing the crude register variables from TI.